pyproject.toml
[project]
name = "river-bank-system"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click (>=8.3.0,<9.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)"
]

[tool.poetry]
packages = [{include = "river_bank_system", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
